added user story one to four
User story 1
As a user i want an overview of a lot of emoji's.

User story 2
As a user i want search for an emojii to find a specific emoji.

User story 3
As a user i want to click on an emoji to copy it to my clipboard.

User story 4
As a user i want to select a category of emoji's.